Moved holypower tracker to shared.
